# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
    "version": "1.0.0",
    "title": "Test API",
    "contact": {
      "email": "interview@ts.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http", "https"],
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
    "/orders": {
      "get": {
        "summary": "Get all the orders",
        "description": "Get all the orders",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "number"
          },
          {
            "in": "query",
            "name": "limit",
            "type": "number"
          },
          {
            "in": "query",
            "name": "sort",
            "description": "only ASC or DESC",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "currentPage": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/ordersResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get One the order",
        "description": "Get One the order",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "order id that needs to be deleted",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/ordersResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ordersResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "customer": {
          "type": "object",
          "properties": {
            "full_name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            }
          }
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        },
        "grand_total": {
          "type": "integer"
        }
      }
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "item_id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "qty": {
          "type": "integer"
        },
        "unit_price": {
          "type": "integer"
        }
      }
    },
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}